requirements.txt
flask==2.3.3
flask-cors
gunicorn==21.2.0
numpy
opencv-python
mediapipe==0.10.9
Pillow